1.0
add scripts for uninstall and config change and change service name

update the gitlab server ip

update scripts

README add description and compatibility

indentation

clone from github

+++ update_discord.sh
@@ -1,31 +1,39 @@
 #!/bin/bash
 
-# Chargement de la config
+# load config
 SCRIPT_DIR="$(dirname "$(readlink -f "$0")")"
-source "$SCRIPT_DIR/config.conf"
 
-# Vérification si un fichier correspond au pattern
-DEB_FILE=$(ls -t $DOWNLOAD_DIR/$FILE_PATTERN 2>/dev/null | head -n 1)
+# Search for the config in ~/.config/adu/ or fallback to the script directory
+if [ -f "$HOME/.config/adu/config.conf" ]; then
+    source "$HOME/.config/adu/config.conf"
+elif [ -f "$SCRIPT_DIR/config.conf" ]; then
+    source "$SCRIPT_DIR/config.conf"
+else
+    echo "❌ Configuration file not found."
+    exit 1
+fi
+
+# Check if a file matches the pattern
+DEB_FILE=$(ls -t "$DOWNLOAD_DIR"/$FILE_PATTERN 2>/dev/null | head -n 1)
 
 if [ -z "$DEB_FILE" ]; then
     echo "❌ Aucun fichier $FILE_PATTERN trouvé dans $DOWNLOAD_DIR"
     exit 1
 fi
 
-echo "🚀 Installation de : $(basename "$DEB_FILE")"
+echo "🚀 Installing: $(basename "$DEB_FILE")"
 
-# Installation (demande le mot de passe sudo)
-sudo dpkg -i "$DEB_FILE"
-
-# Correction des dépendances si nécessaire
-#sudo apt-get install -f -y
-
-# Relancer Discord
-echo "🔄 Redémarrage de Discord..."
-#pkill -9 Discord
-nohup discord > /dev/null 2>&1 &
-
-# Nettoyage : supprimer le .deb après install (optionnel)
-rm "$DEB_FILE"
-
-echo "✅ Mise à jour terminée !"
+# Install the package
+if sudo dpkg -i "$DEB_FILE"; then
+    echo "✅ Installation successful"
+    
+    # Restart Discord
+    echo "🔄 Restarting Discord..."
+    setsid discord > /dev/null 2>&1 &
+    
+    rm "$DEB_FILE"
+    echo "✅ Cleanup complete!"
+else
+    echo "❌ Error during installation. The file has been kept."
+    exit 1
+fi
